+++ src/Ceiling.java
@@ -1,5 +1,5 @@
 public class Ceiling {
-    public static Integer call(Double a) {
-        return (int) Math.ceil(a);
+    public static Double call(Double a) {
+        return Math.ceil(a);
     }
 }
