+++ src/Floor.java
@@ -1,5 +1,5 @@
 public class Floor {
-    public static Integer call(Double a) {
-        return (int) Math.floor(a);
+    public static Double call(Double a) {
+        return Math.floor(a);
     }
 }
